+++ contracts/ProxyReader.sol
@@ -1,7 +1,7 @@
 // @author Unstoppable Domains, Inc.
 // @date June 16th, 2021
 
-pragma solidity ^0.8.0;
+pragma solidity 0.8.17;
 
 import '@openzeppelin/contracts-upgradeable/utils/introspection/ERC165Upgradeable.sol';
 import '@openzeppelin/contracts-upgradeable/utils/MulticallUpgradeable.sol';