+++ CHANGELOG.md
@@ -1,6 +1,17 @@
-## v0.9.14
+## v0.9.16
 
 - Add `.witg` TLD
+- TBD
+
+## v0.9.15
+
+- Add SeaportProxyBuyer contract
+- Update sandbox RPC server to Foundry Anvil to support cancun hardfork
+- Add Seaport contracts to sandbox
+
+## v0.9.14
+
+- Added `scripts/util/deploy_ERC1271SimpleWallet.ts` utility script
 
 ## v0.9.13